refactor(#933): extract events CLI runtime into tau-events
