+++ crates/tau-events/Cargo.toml
@@ -13,6 +13,7 @@ hmac.workspace = true
 serde.workspace = true
 serde_json.workspace = true
 sha2.workspace = true
+tau-cli = { path = "../tau-cli" }
 tau-core = { path = "../tau-core" }
 tau-runtime = { path = "../tau-runtime" }
 tokio.workspace = true
